+++ python/cudf/io/__init__.py
@@ -5,3 +5,4 @@ from cudf.io.parquet import read_parquet
 from cudf.io.feather import read_feather
 from cudf.io.json import read_json
 from cudf.io.hdf import read_hdf
+from cudf.io.orc import read_orc